% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_crosses_in_R.R
\name{do_pun}
\alias{do_pun}
\title{Function to create punnett square given the genotypes of parents}
\usage{
do_pun(female.geno, male.geno)
}
\arguments{
\item{female.geno}{is a character object for the genotype of the female
parent.}

\item{male.geno}{is a character object for the genotype of the male parent.}
}
\value{
a punnett square table showing the genotypes of progenies in a
genetic cross.
}
\description{
Function to create punnett square given the genotypes of parents
}
\examples{
# library(bemendel)
pun1 <- do_pun(female.geno = "AaBbDd", male.geno = "AaBbDd")
pun1$punnett_square
pun1$pun_summary
}
