% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_crosses_in_R.R
\name{epist}
\alias{epist}
\title{Function to display classical digenic epistatic ratios. Five types are
available}
\usage{
epist(female.geno, male.geno, type = c("DDE", "DRE", "DE", "RE", "DnRE"))
}
\arguments{
\item{female.geno}{is a character object for the genotype of the female
parent.}

\item{male.geno}{is a character object for the genotype of the male parent.}

\item{type}{a string abbreviation indicating the type of digenic epistasis;
see details for more information.}
}
\value{
a melted long format data frame.
}
\description{
Function to display classical digenic epistatic ratios. Five types are
available
}
\details{
The following classical digenic epistasis types are implemented:
DDE: Duplicate dominant epistasis (15:1 ratio)
DRE: Duplicate recessive epistasis (9:7 ratio)
DE: Dominant epistasis (12:3:1 ratio)
RE: Recessive epistasis (9:3:4 ratio)
DnRE: Dominant and recessive epistasis (13:3 ratio)
}
\examples{
# library(bemendel)
epi_de <- epist(female.geno = "AaBb", male.geno = "AaBb", type = "DE")
epi_de
table(epi_de$epistasis)
}
