% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_crosses_in_R.R
\name{gamete}
\alias{gamete}
\title{Function to generate parental gametes}
\usage{
gamete(x)
}
\arguments{
\item{x}{a list object containing split loci and alleles generated by
the \code{split_geno()} function}
}
\value{
a factor object that contain all possible gametes that can be
produced by an individual at meiosis.
}
\description{
Function to generate parental gametes
}
\examples{
 # library(bemendel)
 mat_geno <- split_geno("AaBb")$Loci
 mat_gam <- gamete(mat_geno)
 mat_gam
}
