% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_crosses_in_R.R
\name{is.het}
\alias{is.het}
\title{Function to check if locus is heterozygous in diploid species}
\usage{
is.het(x)
}
\arguments{
\item{x}{a character vector of split alleles for each locus--
output of the \code{split_geno()} function.}
}
\value{
a logical value of TRUE if locus is heterozygous or FALSE if locus is
homozygous.
}
\description{
Function to check if locus is heterozygous in diploid species
}
\examples{
# library(bemendel)
is.het(split_geno("AaBb")$Loci[[1]])
}
