% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_crosses_in_R.R
\name{is.seg}
\alias{is.seg}
\title{Check whether a locus will segregate using this function}
\usage{
is.seg(x)
}
\arguments{
\item{x}{a character vector object of split alleles for each locus.}
}
\value{
returns a string about whether locus is homozygous or heterozygous.
}
\description{
Check whether a locus will segregate using this function
}
\examples{
# Output is a string object
x <- split_geno("AABb")$Loci
is.seg(x[[1]])
is.seg(x[[2]])
}
