% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_crosses_in_R.R
\name{order_als}
\alias{order_als}
\title{Function to order allele combinations at a locus -- Dominant allele first
Order alleles (dominant allele comes first at each locus)}
\usage{
order_als(x)
}
\arguments{
\item{x}{a list object containing split loci and alleles at each locus.}
}
\value{
ordered alleles at each locus.
}
\description{
Function to order allele combinations at a locus -- Dominant allele first
Order alleles (dominant allele comes first at each locus)
}
\examples{
 # library(bemendel)
 mat_geno <- split_geno("AaBb")$Loci
 order_als(mat_geno$Locus_1)
}
