% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_crosses_in_R.R
\name{pun2df}
\alias{pun2df}
\title{Function to melt punnett square to a long format data frame for plotting}
\usage{
pun2df(pun)
}
\arguments{
\item{pun}{a matrix of punnett square containing progeny genotypes.}
}
\value{
a long format data frame for plotting punnett square.
}
\description{
Function to melt punnett square to a long format data frame for plotting
}
\examples{
# library(bemendel)
pun1 <- do_pun(female.geno = "AaBb", male.geno = "AaBb")
long_df <- pun2df(pun1$punnett_square)
long_df
}
