% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_crosses_in_R.R
\name{pun_plot}
\alias{pun_plot}
\title{Plot Punnett square to visualize phenotype classes of progenies assuming
complete dominance at each locus. The function helps users to visualize
predictions for common gene interaction hypotheses such as independent
assortment and classical digenic interactions.}
\usage{
pun_plot(pun2df, text_size = 12, epistasis = FALSE)
}
\arguments{
\item{pun2df}{a melted long format data frame of punnett square containing
progeny genotypes and phenotypic class designation.}

\item{text_size}{a numeric value for setting text size in plot.}

\item{epistasis}{a logical value indicating whether input data frame contains
digenic epistatic phenotypes.}
}
\value{
a plot of the punnett square showing the phenotypic classes for
gene action hypothesis.
}
\description{
Plot Punnett square to visualize phenotype classes of progenies assuming
complete dominance at each locus. The function helps users to visualize
predictions for common gene interaction hypotheses such as independent
assortment and classical digenic interactions.
}
\details{
Predictions of phenotypic classes for independent assortment are based on
complete dominance at all gene loci. For epistasis, predictions are based on
intergenic allelic interactions between two loci.

For epistasis, the number of loci must be equal to 2. Digenic epistatic
ratios that can be visualized are
Duplicate dominant epistasis = 15:1 (Duplicate gene action)
Duplicate recessive epistasis = 9:7 (Complementary gene action)
Dominant epistasis = 12:3:1
Recessive epistasis = 9:3:4
Dominant and recessive epistasis = 13:3 (recessive suppressors)
}
\examples{
# library(bemendel)
# Generate punnett square
pun1 <- do_pun(female.geno = "AaBb", male.geno = "AaBb")

# Melt punnett to a long format data frame for plotting
long_df <- pun2df(pun1$punnett_square)

# Visualize phenotypic groups based on independent assortment
pun_plot(pun2df = long_df, text_size = 6)

# Visualize phenotypic groups based on dominant epistatic gene action
epi_de <- epist(female.geno = "AaBb", male.geno = "AaBb", type = "DE")
pun_plot(pun2df = epi_de, text_size = 6, epistasis = TRUE)
}
