% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the bemendel Shiny app.}
\usage{
run_app()
}
\description{
An interface function to launch a user-friendly Shiny app for
making genetic crosses in R.
}
\concept{"Mendelian genetics in R"}
