% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_crosses_in_R.R
\name{split_geno}
\alias{split_geno}
\title{Function to split inputted parental genotypes into alleles for each locus in
a diploid species}
\usage{
split_geno(x)
}
\arguments{
\item{x}{is a character or string object of genotype for at least one locus.}
}
\value{
a list object of two elements representing the number of loci and
split loci and loci alleles
}
\description{
Function to split inputted parental genotypes into alleles for each locus in
a diploid species
}
\details{
\code{x} should be either lower or uppercase letters of the Latin/English alphabet.
Uppercase letters represent dominant alleles, and lowercase letters denote
recessive alleles.
}
\examples{
# library(bemendel)
mat_geno <- split_geno("AaBb")$Loci
mat_geno
}
